# server/docs/swagger.yaml
components:
  schemas:
    api.ErrorResponse:
      properties:
        code:
          example: 400
          type: integer
        error:
          example: validation failed
          type: string
        message:
          example: Bad request
          type: string
      type: object
...
    dto.LocationClusterDTO:
      properties:
        centroid_latitude:
          example: 37.7749
          type: number
        centroid_longitude:
          example: -122.4194
          type: number
        city:
          example: San Francisco
          type: string
        cluster_id:
          example: 550e8400-e29b-41d4-a716-446655440000
          type: string
        country:
          example: United States
          type: string
        geocode_status:
          example: resolved
          type: string
        geocoded_at:
          example: "2026-02-10T12:00:00Z"
          type: string
        geohash:
          example: 9q8yyk8
          type: string
        label:
          example: San Francisco, California, United States
          type: string
        photo_count:
          example: 42
          type: integer
        precision:
          example: 7
          type: integer
        provider:
          example: nominatim
          type: string
        region:
          example: California
          type: string
        repository_id:
          example: 550e8400-e29b-41d4-a716-446655440000
          type: string
      type: object
    dto.LocationClusterListResponseDTO:
      properties:
        clusters:
          items:
            $ref: '#/components/schemas/dto.LocationClusterDTO'
          type: array
          uniqueItems: false
        limit:
          example: 100
          type: integer
        offset:
          example: 0
          type: integer
        total:
          example: 150
          type: integer
      type: object
...
    dto.RebuildLocationClustersRequestDTO:
      properties:
        repository_id:
          example: 550e8400-e29b-41d4-a716-446655440000
          type: string
      type: object
    dto.RebuildLocationClustersResponseDTO:
      properties:
        job_id:
          example: 123
          type: integer
        message:
          example: Location cluster rebuild queued successfully
          type: string
        repository_id:
          example: 550e8400-e29b-41d4-a716-446655440000
          type: string
        status:
          example: queued
          type: string
      type: object
...
  securitySchemes:
    BearerAuth:
      description: Type "Bearer" followed by a space and JWT token.
      in: header
      name: Authorization
      type: apiKey
...
info:
  contact:
    name: API Support
    url: http://www.github.com/EdwinZhanCN/Lumilio-Photos
  description: Media management system API with asset features
  license:
    name: GPLv3.0
    url: https://opensource.org/licenses/GPL-3.0
  title: Lumilio-Photos API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /api/v1/locations/clusters:
    get:
      description: Return paginated persisted photo location clusters with cached
        labels when available.
      parameters:
      - description: Page size (1-1000)
        in: query
        name: limit
        schema:
          default: 100
          type: integer
      - description: Page offset
        in: query
        name: offset
        schema:
          default: 0
          type: integer
      - description: Optional repository UUID filter
        in: query
        name: repository_id
        schema:
          type: string
      - description: Optional geohash filter
        in: query
        name: geohash
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.LocationClusterListResponseDTO'
          description: Location clusters retrieved successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Invalid request parameters
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal server error
      summary: Get location clusters
      tags:
      - locations
  /api/v1/locations/rebuild:
    post:
      description: Queue a location cluster rebuild for all photos or one repository.
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/dto.RebuildLocationClustersRequestDTO'
                description: Rebuild request
                summary: request
        description: Rebuild request
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.RebuildLocationClustersResponseDTO'
          description: Location cluster rebuild queued successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Invalid request body
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal server error
      summary: Queue location cluster rebuild
      tags:
      - locations
...
servers:
- url: localhost:6680/api/v1